#2281 [BE] Design and Implement getting contributors from our github project, save them into our DB, prepare api for getting them to our front * Moved enum to enum package
