+++ src/main/kotlin/com/epam/brn/enums/ContactType.kt
@@ -1,4 +1,4 @@
-package com.epam.brn.model
+package com.epam.brn.enums
 
 enum class ContactType {
     EMAIL,
