+++ src/main/kotlin/com/epam/brn/service/GitHubContributorService.kt
@@ -1,7 +1,7 @@
 package com.epam.brn.service
 
 import com.epam.brn.dto.ContributorUserDto
-import com.epam.brn.model.ContributorType
+import com.epam.brn.enums.ContributorType
 import com.epam.brn.repo.ContributorRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
